Disable docs when building jq
It doens't make much of a difference on 1.6, but on master it avoids
having to install pipenv dependencies, which, otherwise, result in
a build error. The jq configure should probably check for the pipenv
dependencies instead of assuming the presence of pipenv means they'll
be installed (in a separate step from the make), but there's no point
in building with docs in this context anyway.